% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_numeric_vector}
\alias{check_numeric_vector}
\title{Validate a numeric vector}
\usage{
check_numeric_vector(x, name)
}
\arguments{
\item{x}{\strong{numeric} vector to validate.}

\item{name}{\strong{character} string used in error messages.}
}
\value{
Invisible \code{TRUE} when validation succeeds.
}
\description{
Ensures an argument is a numeric vector with no missing or infinite values.
}
\keyword{internal}
