% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_blockchain_audit.R
\name{pharma_audit_get}
\alias{pharma_audit_get}
\title{Retrieve the audit blockchain}
\usage{
pharma_audit_get()
}
\value{
A \code{data.frame} of blockchain entries.
}
\description{
Returns a data frame representing the current audit chain.
}
