% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_blockchain_audit.R
\name{pharma_audit_init}
\alias{pharma_audit_init}
\title{Initialize or reset audit blockchain}
\usage{
pharma_audit_init()
}
\value{
Invisible \code{NULL}.
}
\description{
Creates an empty blockchain ledger for storing analysis audit information.
}
