% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_bayesian.R
\name{pharma_bayesian_glm}
\alias{pharma_bayesian_glm}
\title{Bayesian GLM with rstanarm}
\usage{
pharma_bayesian_glm(formula, data, ...)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame}

\item{...}{Additional arguments passed to \link[rstanarm:stan_glm]{rstanarm::stan_glm}}
}
\value{
A \code{stanreg} object
}
\description{
Fit a generalized linear model using MCMC via \code{rstanarm::stan_glm}.
}
\examples{
\dontrun{
  fit <- pharma_bayesian_glm(outcome ~ dose, data = pharma_sample, iter = 500)
}
}
