% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_block_bootstrap.R
\name{pharma_block_bootstrap}
\alias{pharma_block_bootstrap}
\title{Block bootstrap for clustered data}
\usage{
pharma_block_bootstrap(data, cluster, statistic, R = 1000, ...)
}
\arguments{
  \item{data}{Data frame to resample.}
  \item{cluster}{Vector or column name indicating cluster membership.}
  \item{statistic}{Function computing the statistic of interest.}
  \item{R}{Number of bootstrap replicates.}
  \item{...}{Additional arguments passed to \code{statistic}.}
}
\value{A list of bootstrap statistics.}
\description{Resamples entire clusters with replacement to obtain bootstrap statistics.}
\examples{
stat <- function(d) coef(lm(response ~ treatment, data = d))[2]
res <- pharma_block_bootstrap(pharma_sample, "subject", stat, R = 10)
}
