% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_greeting.R
\name{pharma_greeting}
\alias{pharma_greeting}
\title{Pharma greeting}
\usage{
pharma_greeting()
}
\value{
A character string with the greeting message.
}
\description{
This function returns a friendly greeting for PharmaStatsR users.
}
