% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_plugin_api.R
\name{pharma_list_plugins}
\alias{pharma_list_plugins}
\title{List registered PharmaTestSuite plugins}
\usage{
pharma_list_plugins()
}
\value{
Character vector of plugin names.
}
\description{
List registered PharmaTestSuite plugins
}
\examples{
pharma_list_plugins()
}
