% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_plugin_api.R
\name{pharma_list_stats}
\alias{pharma_list_stats}
\title{List registered statistical methods}
\usage{
pharma_list_stats()
}
\value{
Character vector of registered method names.
}
\description{
List registered statistical methods
}
\examples{
pharma_list_stats()
}
