% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pharma_log}
\alias{pharma_log}
\title{Log package events}
\usage{
pharma_log(level = "INFO", msg)
}
\arguments{
\item{level}{**character** log level for the message.}

\item{msg}{**character** string to output.}
}
\value{
Invisible `NULL`.
}
\description{
Simple logging utility with verbosity controlled by
`pharma_config()` via the `log_level` option. Levels in increasing
order are "DEBUG", "INFO", "WARN", and "ERROR".
}
\keyword{internal}
