% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_mice_impute.R
\name{pharma_mice_impute}
\alias{pharma_mice_impute}
\title{Perform multiple imputation using mice}
\usage{
pharma_mice_impute(data, m = 5, ...)
}
\arguments{
\item{data}{A data frame with missing values.}

\item{m}{Number of imputations to perform. Defaults to 5.}

\item{...}{Additional arguments passed to `mice::mice`.}
}
\value{
A `mids` object from the **mice** package.
}
\description{
Convenience wrapper around `mice::mice` for creating multiply imputed
datasets. Requires the optional **mice** package.
}
\examples{
if (requireNamespace("mice", quietly = TRUE)) {
  dat <- pharma_sample
  dat$response[1] <- NA
  pharma_mice_impute(dat, m = 2, maxit = 1)
}
}
