% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_multistate_model.R
\name{pharma_multistate_model}
\alias{pharma_multistate_model}
\title{Fit a multistate illness-death model}
\usage{
pharma_multistate_model(coxFit, trans, ...)
}
\arguments{
\item{coxFit}{A Cox model fitted on data prepared with `mstate::msprep`.}

\item{trans}{A transition matrix created by `mstate::transMat`.}

\item{...}{Additional arguments passed to `mstate::msfit`.}
}
\value{
An object of class `msfit`.
}
\description{
Wrapper around `mstate::msfit` for estimating transition-specific
hazards in multistate analyses.
}
\examples{
# Fit a Cox model on data prepared with mstate::msprep first.
}
