% Generated by roxygen2
% Please edit documentation in R/pharma_bayesian.R
\name{pharma_pp_check}
\alias{pharma_pp_check}
\title{Posterior predictive check}
\usage{
pharma_pp_check(model, ...)
}
\arguments{
  \item{model}{A \code{stanreg} object}
  \item{...}{Additional arguments passed to \code{bayesplot::pp_check}}
}
\value{
A \code{ggplot} object
}
\description{
Generate a posterior predictive check plot for a fitted Bayesian model using \code{bayesplot::pp_check}.
}
\seealso{\code{\link[bayesplot]{pp_check}}}
