% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_plugin_api.R
\name{pharma_register_stat}
\alias{pharma_register_stat}
\title{Register a custom statistical method}
\description{
Extend PharmaStatsR with bespoke statistical routines that can be
invoked through \code{pharma_run_stat()}.
}
\usage{
pharma_register_stat(name, fun)
}
\arguments{
\item{name}{Name of the method.}
\item{fun}{Function implementing the method.}
}
\value{
Invisible \code{TRUE} on success.
}
\examples{
pharma_register_stat("mean_method", mean)
pharma_run_stat("mean_method", 1:3)
pharma_unregister_stat("mean_method")
}
