% Generated by roxygen2
\name{pharma_tost}
\alias{pharma_tost}
\title{Two One-Sided Tests (TOST) for Equivalence}
\usage{
pharma_tost(x, y, low_eqbound, high_eqbound, alpha = 0.05)
}
\arguments{
  \item{x}{Numeric vector of group 1 observations.}
  \item{y}{Numeric vector of group 2 observations.}
  \item{low_eqbound}{Lower equivalence bound.}
  \item{high_eqbound}{Upper equivalence bound.}
  \item{alpha}{Significance level.}
}
\value{A list with test statistics, p-value, difference and confidence interval.}
\description{Perform a two one-sided t-test to assess equivalence of two means.}
\examples{
pharma_tost(rnorm(30), rnorm(30, 0.1), -0.5, 0.5)
}
