% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_plugin_api.R
\name{pharma_unregister_stat}
\alias{pharma_unregister_stat}
\title{Unregister a statistical method}
\usage{
pharma_unregister_stat(name)
}
\arguments{
\item{name}{Name of the statistical method to remove.}
}
\value{
Invisible \code{TRUE} whether or not the method existed.
}
\description{
Unregister a statistical method
}
