% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pharma_dose_response.R
\name{pharma_dose_response}
\alias{pharma_dose_response}
\title{Simulated dose-response dataset}
\format{A data frame with 30 rows and 3 variables:}
\describe{
  \item{subject}{Subject identifier}
  \item{dose}{Dose level}
  \item{response}{Observed response}
}
\source{Simulated data.}
