% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_meta_analysis.R
\name{pharma_funnel_plot}
\alias{pharma_funnel_plot}
\title{Funnel plot for a meta-analysis}
\usage{
pharma_funnel_plot(model, ...)
}
\arguments{
  \item{model}{An object from \code{metafor::rma}.}
  \item{...}{Additional arguments passed to \code{metafor::funnel}.}
}
\value{
A plot.
}
\description{
Create a funnel plot from an \code{rma} object.
}
\examples{
res <- pharma_meta_analysis(yi = c(0.2,0.1,-0.1), vi=c(0.05,0.04,0.06))
pharma_funnel_plot(res)
}
