% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_parallel_bootstrap.R
\name{pharma_parallel_bootstrap}
\alias{pharma_parallel_bootstrap}
\title{Parallel bootstrap using future.apply}
\usage{
pharma_parallel_bootstrap(data, statistic, R = 1000, plan = "multisession", ...)
}
\arguments{
  \item{data}{Data frame containing the sample.}
  \item{statistic}{Function computing the statistic of interest.}
  \item{R}{Number of bootstrap replicates.}
  \item{plan}{Future plan specification.}
  \item{...}{Additional arguments passed to \code{statistic}.}
}
\value{A list of bootstrap statistics.}
\description{Distribute bootstrap iterations across multiple cores using the \code{future} framework.}
\examples{
if (requireNamespace("future.apply", quietly = TRUE)) {
stat <- function(d) mean(d$response)
pharma_parallel_bootstrap(pharma_sample, stat, R = 10, plan = "sequential")
}
}
