% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_bayesian.R
\name{pharma_pp_check}
\alias{pharma_pp_check}
\title{Posterior predictive check}
\usage{
pharma_pp_check(model, ...)
}
\arguments{
\item{model}{A `stanreg` object}

\item{...}{Additional arguments passed to [bayesplot::pp_check]}
}
\value{
A `ggplot` object
}
\description{
Generate a posterior predictive check plot for a fitted Bayesian model
using [bayesplot::pp_check].
}
\examples{
\dontrun{
  fit <- pharma_bayesian_glm(outcome ~ dose, data = pharma_sample, iter = 500)
  pharma_pp_check(fit)
}
}
