% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pharma_progress}
\alias{pharma_progress}
\title{Display progress for long-running operations}
\usage{
pharma_progress(max_value, msg = "Progress")
}
\arguments{
\item{max_value}{**integer** total number of steps.}

\item{msg}{**character** base message to display with progress updates.}
}
\value{
A list with `update` and `finish` functions.
}
\description{
Creates a simple progress reporter that outputs status messages to the
console when running interactively.
}
\keyword{internal}
