% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pharma_repeated.R
\name{pharma_repeated}
\alias{pharma_repeated}
\title{Simulated repeated measures dataset}
\format{A data frame with 20 rows and 3 variables:
\describe{
  \item{subject}{Subject identifier}
  \item{condition}{Condition ("A" or "B")}
  \item{response}{Numeric response}
}}
\source{Simulated data.}
