% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_repeated_anova.R
\name{pharma_repeated_anova}
\alias{pharma_repeated_anova}
\title{Conduct a repeated measures ANOVA}
\usage{
pharma_repeated_anova(formula, data, ...)
}
\arguments{
  \item{formula}{Model formula specifying within-subject error terms.}
  \item{data}{Data frame containing the variables used in the model.}
  \item{...}{Additional arguments passed to \code{aov}.}
}
\value{An object of class \code{aov}.}
\description{Wrapper around \code{stats::aov} for repeated measures designs.}
\examples{pharma_repeated_anova(response ~ condition + Error(subject), data = pharma_repeated)}
