% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_plugin_api.R
\name{pharma_unregister_stat}
\alias{pharma_unregister_stat}
\title{Unregister a statistical method}
\usage{
pharma_unregister_stat(name)
}
\arguments{
\item{name}{Character string of the plugin to remove.}
}
\value{
Invisible `TRUE` whether or not the method existed.
}
\description{
Unregister a statistical method
}
