% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_wild_bootstrap.R
\name{pharma_wild_bootstrap}
\alias{pharma_wild_bootstrap}
\title{Wild bootstrap for linear models}
\usage{
pharma_wild_bootstrap(formula, data, R = 1000)
}
\arguments{
  \item{formula}{Model formula passed to \code{lm}.}
  \item{data}{Data frame containing model variables.}
  \item{R}{Number of bootstrap replicates.}
}
\value{A matrix of bootstrap coefficients.}
\description{Applies the Rademacher wild bootstrap to a linear model.}
\examples{
res <- pharma_wild_bootstrap(response ~ treatment, data = pharma_sample, R = 10)
head(res)
}
