% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{with_pharma_config}
\alias{with_pharma_config}
\title{Temporarily modify configuration options}
\usage{
with_pharma_config(options, expr)
}
\arguments{
\item{options}{Named list of options to set.}

\item{expr}{Expression to evaluate while the options are active.}
}
\value{
The result of `expr`.
}
\description{
`with_pharma_config()` sets package options for the duration of `expr` and
restores the previous values on exit.
}
\examples{
with_pharma_config(list(log_level = "DEBUG"), {
  pharma_config()$log_level
})
}
